fix: split Mermaid viewport helpers to satisfy lint budget
Extract fit/pan and SVG sizing helpers so max-lines-per-function and complexity stay within CI budgets.
